prime factors

+++ 0x04-more_functions_nested_loops/100-prime_factor.c
@@ -8,17 +8,17 @@
  */
 int main(void)
 {
-	int count;
+	long int count;
 
 	long int num = 612852475143;
 
-	for (count = 2; num > count;  count++)
+	for (count = 2; num > count; count++)
 	{
-		while (num % count ==0)
+		while (num % count == 0)
 		{
 			num = num / count;
 		}
 	}
-	printf("%d\n", count);
+	printf("%ld\n", count);
 	return (0);
 }
